requirements.txt
mysql-connector
PyQt5